fix: 修复 PR #8 二轮审查问题
🔴 阻断性问题:
- user_data_store 改为 JSONL 文件持久化 + 线程锁，解决重启丢数据和并发安全问题

🟡 建议项:
- 移除 Workbench mount 时的重复 onRefresh() 调用，避免双重 API 请求
- 算法切换添加 300ms debounce，防止快速点击冲击服务器
- 非树模型不确定性从硬编码 0.0 改为训练残差标准差
- Workbench 简化为直接使用 props，移除内部 state 镜像冗余

💭 细节:
- marking_count 类型从 float 改为 int（语义正确性）
- clean_raw_data.py 回填配置从索引号改为列名引用，降低新增数据集时出错风险

+++ apps/api/app/schemas.py
@@ -103,7 +103,7 @@ class ExperimentData(BaseModel):
     pulse_energy_mj: float | None = Field(default=None, ge=0, le=1000)
     laser_energy_percent: float | None = Field(default=None, ge=0, le=100)
     defocus_amount_mm: float | None = Field(default=None, ge=-100, le=100)
-    marking_count: float | None = Field(default=None, ge=0, le=1000)
+    marking_count: int | None = Field(default=None, ge=0, le=1000)
     fill_spacing_um: float | None = Field(default=None, ge=0, le=10000)
     scan_interval_um: float | None = Field(default=None, ge=0, le=100000)
     processing_time_s: float | None = Field(default=None, ge=0, le=360000)
